% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA_Mar.R
\name{CA_Mar}
\alias{CA_Mar}
\title{CA_Mar}
\usage{
CA_Mar(scratch_env = NULL, HL = NULL)
}
\arguments{
\item{scratch_env}{default is \code{NULL} This is an environment containing the results of a
Mar.datawrangling extraction - i.e. it contains all of the data necessary for HH, HL and CA}
}
\value{
a df generated CA file
}
\description{
This function generates ICES DATRAS-compatible "CA" files 
directly from the Maritimes. "CA" files contain Species age-based information
}
\seealso{
Other DATRAS: 
\code{\link{DATRAS_Cen}()},
\code{\link{DATRAS_Gulf}()},
\code{\link{DATRAS_Mar}()},
\code{\link{DATRAS_NFLD}()},
\code{\link{DATRAS_Pac}()},
\code{\link{DATRAS_Que}()},
\code{\link{HH_Mar}()},
\code{\link{HL_Mar}()},
\code{\link{makeDATRAS}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{DATRAS}
