% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATRAS_Gulf.R
\name{DATRAS_Gulf}
\alias{DATRAS_Gulf}
\title{DATRAS_Gulf}
\usage{
DATRAS_Gulf(yr = NULL, season = NULL, csv = NULL)
}
\arguments{
\item{yr}{default is \code{NULL}. This specifies the year(s) for which you'd like to generate
HH files. Single years are fine, as are vectors (e.g. \code{c(2011,1015)}, \code{2015:2019})}

\item{season}{default is \code{NULL}.}

\item{csv}{default is \code{TRUE}.  If \code{TRUE}, csv files are generated for each HH code.  If
\code{FALSE}, the output exists only in the resultant list.}
}
\value{
a list containing (named) objects - 1 for each generated HH file
}
\description{
This is a placeholder for an imaginary function which would generate
ICES DATRAS-compatible files for Gulf region.
}
\seealso{
Other DATRAS: 
\code{\link{CA_Mar}()},
\code{\link{DATRAS_Cen}()},
\code{\link{DATRAS_Mar}()},
\code{\link{DATRAS_NFLD}()},
\code{\link{DATRAS_Pac}()},
\code{\link{DATRAS_Que}()},
\code{\link{HH_Mar}()},
\code{\link{HL_Mar}()},
\code{\link{makeDATRAS}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{DATRAS}
