% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HH_Mar.R
\name{HH_Mar}
\alias{HH_Mar}
\title{HH_Mar}
\usage{
HH_Mar(scratch_env = NULL)
}
\arguments{
\item{scratch_env}{default is \code{NULL} This is an environment containing the results of a
Mar.datawrangling extraction - i.e. it contains all of the data necessary for HH, HL and CA}
}
\value{
a df generated HH file
}
\description{
This function generates ICES DATRAS-compatible "HH" files 
directly from the Maritimes groundfish database. "HH" files contain Haul 
metadata
}
\seealso{
Other DATRAS: 
\code{\link{CA_Mar}()},
\code{\link{DATRAS_Cen}()},
\code{\link{DATRAS_Gulf}()},
\code{\link{DATRAS_Mar}()},
\code{\link{DATRAS_NFLD}()},
\code{\link{DATRAS_Pac}()},
\code{\link{DATRAS_Que}()},
\code{\link{HL_Mar}()},
\code{\link{makeDATRAS}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
 @export
drop unneeded fields, and make original cols lowercase - finals will be uppercase
we record the direction of the current relative to the ship (eg to bow,
to starboard, etc), but ICES wants it in a compass direction.
I use the tow direction (determined above) combined with the
value stored in GSINF.CURNT  to derive the value.
For example, if the tow was found to be West to East (90deg), and the CURNT value was
"to starboard", the combined angle would be 90+90 = 180 (or South)
I am awaiting direction on the values below from folks who are familiar with the gear
In the meantime all of these become "-9"
}
\concept{DATRAS}
