% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HL_Mar.R
\name{HL_Mar}
\alias{HL_Mar}
\title{HL_Mar}
\usage{
HL_Mar(scratch_env = NULL)
}
\arguments{
\item{scratch_env}{default is \code{NULL} This is an environment containing the results of a
Mar.datawrangling extraction - i.e. it contains all of the data necessary for HH, HL and CA}
}
\value{
a df generated HL file
}
\description{
This function generates ICES DATRAS-compatible HL files directly from the Maritimes
groundfish database.
}
\seealso{
Other DATRAS: 
\code{\link{DATRAS_Cen}()},
\code{\link{DATRAS_Gulf}()},
\code{\link{DATRAS_Mar}()},
\code{\link{DATRAS_NFLD}()},
\code{\link{DATRAS_Pac}()},
\code{\link{DATRAS_Que}()},
\code{\link{HH_Mar}()},
\code{\link{makeDATRAS}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{DATRAS}
