% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mar_CA.R
\name{Mar_CA}
\alias{Mar_CA}
\title{Mar_CA}
\usage{
Mar_CA(scratch_env = NULL)
}
\arguments{
\item{scratch_env}{default is \code{NULL} This is an environment containing the results of a
Mar.datawrangling extraction - i.e. it contains all of the data necessary for HH, HL and CA}
}
\value{
a df generated CA file
}
\description{
This function generates ICES DATRAS-compatible "CA" files 
directly from the Maritimes. "CA" files contain Species age-based information
}
\seealso{
Other DATRAS: 
\code{\link{Cen_DATRAS}()},
\code{\link{Gulf_DATRAS}()},
\code{\link{Mar_DATRAS}()},
\code{\link{Mar_HH}()},
\code{\link{Mar_HL}()},
\code{\link{NFLD_DATRAS}()},
\code{\link{Pac_DATRAS}()},
\code{\link{Que_DATRAS}()},
\code{\link{makeDATRAS}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{DATRAS}
