% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Que_DATRAS.R
\name{Que_DATRAS}
\alias{Que_DATRAS}
\title{Que_DATRAS}
\usage{
Que_DATRAS(yr = NULL, season = NULL, csv = NULL, debug = FALSE)
}
\arguments{
\item{yr}{default is \code{NULL}. This specifies the year(s) for which you'd like to generate
HH files. Single years are fine, as are vectors (e.g. \code{c(2011,1015)}, \code{2015:2019})}

\item{season}{default is \code{NULL}.}

\item{csv}{default is \code{TRUE}.  If \code{TRUE}, csv files are generated for each HH code.  If
\code{FALSE}, the output exists only in the resultant list.}

\item{debug}{The default is \code{F}. Setting this to TRUE will limit the 
results to a single set for a single species.}
}
\value{
a list containing (named) objects - 1 for each generated HH file
}
\description{
This is a placeholder for an imaginary function which would generate
ICES DATRAS-compatible files for Quebec region.
}
\seealso{
Other DATRAS: 
\code{\link{Cen_DATRAS}()},
\code{\link{Gulf_DATRAS}()},
\code{\link{Mar_CA}()},
\code{\link{Mar_DATRAS}()},
\code{\link{Mar_HH}()},
\code{\link{Mar_HL}()},
\code{\link{NFLD_DATRAS}()},
\code{\link{Pac_DATRAS}()},
\code{\link{makeDATRAS}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{DATRAS}
