% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDATRAS.R
\name{makeDATRAS}
\alias{makeDATRAS}
\title{makeDATRAS}
\usage{
makeDATRAS(
  region = NULL,
  yr = NULL,
  survey = NULL,
  csv = T,
  fn.oracle.username = "_none_",
  fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_",
  usepkg = "rodbc",
  data.dir = NULL,
  debug = F
)
}
\arguments{
\item{region}{default is \code{NULL}. This is the Canadian DFO region for which you want to
make the DATRAS files.  Valid regions are listed below, but initially, only "Mar" works (Jan, 2020)
\itemize{
\item \code{Mar}
\item \code{NFLD}
\item \code{Gulf}
\item \code{Que}
\item \code{Cen}
\item \code{Pac}
}}

\item{yr}{default is \code{NULL}. This specifies the year(s) for which you'd like to generate
HH files. Single years are fine, as are vectors (e.g. \code{c(2011,1015)}, \code{2015:2019})}

\item{survey}{default is \code{NULL}. This specifies the survey for which you'd like to generate
HH files.  This will be specific to the different regions.  Valid values for Maritimes are
\itemize{
\item \code{4X} - Type 1; Spring (i.e. months 1:4); 2008+; not strata 5Z* 
\item \code{GEORGES} - Type 1; Spring (i.e. months 1:4);  2008+; strata 5Z*
\item \code{SPRING} - Type 1; Spring (i.e. months 1:4); pre-2008; specific strata 
\item \code{4VSW}  - Type 1; Spring (i.e. months 1:4); 4VSW strata;  
\item \code{SUMMER} - Type 1; Summer (i.e. months 5:8); specific strata
\item \code{FALL} - Type 1; Fall (i.e. months 9:12)
}}

\item{csv}{default is \code{TRUE}.  If \code{TRUE}, csv files are generated for each HH code.
If \code{FALSE}, the output exists only in the resultant list.}

\item{fn.oracle.username}{default is \code{'_none_'} Required for Maritimes HH files. This is
your username for accessing oracle objects. If you have a value for \code{oracle.username}
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take
priority over your existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} Required for Maritimes HH files. This is
your password for accessing oracle objects. If you have a value for \code{oracle.password}
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take
priority over your existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} Required for Maritimes HH files. This is your
dsn/ODBC identifier for accessing oracle objects. If you have a value for
\code{oracle.dsn} stored in your environment (e.g. from an rprofile file),
this can be left and that value will be used.  If a value for this is
provided, it will take priority over your existing value.}

\item{usepkg}{default is \code{'rodbc'}. Required for Maritimes HH files. This indicates whether
the connection to Oracle should use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is
slightly easier to setup, but
roracle will extract data ~ 5x faster.}

\item{data.dir}{The default is \code{NULL}. Required for Maritimes HH files. This is the path
to your Mar.datawrangling rdata files}

\item{debug}{The default is \code{F}. Setting this to TRUE will limit the 
results to a single set for a single species.}
}
\value{
a list containing (named) objects - 1 for each generated HH file
}
\description{
This function generates ICES DATRAS-compatible HH for a particular DFO region.
}
\seealso{
Other DATRAS: 
\code{\link{Cen_DATRAS}()},
\code{\link{Gulf_DATRAS}()},
\code{\link{Mar_CA}()},
\code{\link{Mar_DATRAS}()},
\code{\link{Mar_HH}()},
\code{\link{Mar_HL}()},
\code{\link{NFLD_DATRAS}()},
\code{\link{Pac_DATRAS}()},
\code{\link{Que_DATRAS}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{DATRAS}
